Move utils from models.utils to utils folder
