+++ llmfoundry/models/utils/__init__.py
@@ -3,8 +3,6 @@
 
 from llmfoundry.models.utils.adapt_tokenizer import (
     AutoTokenizerForMOD, adapt_tokenizer_for_denoising)
-from llmfoundry.models.utils.checkpoint_conversion_helpers import (
-    convert_and_save_ft_weights, get_hf_tokenizer_from_composer_state_dict)
 from llmfoundry.models.utils.hf_prefixlm_converter import (
     add_bidirectional_mask_if_missing, convert_hf_causal_lm_to_prefix_lm)
 from llmfoundry.models.utils.meta_init_context import (init_empty_weights,
@@ -15,9 +13,7 @@ from llmfoundry.models.utils.param_init_fns import (  # type: ignore
 __all__ = [
     'AutoTokenizerForMOD',
     'adapt_tokenizer_for_denoising',
-    'convert_and_save_ft_weights',
     'convert_hf_causal_lm_to_prefix_lm',
-    'get_hf_tokenizer_from_composer_state_dict',
     'init_empty_weights',
     'init_on_device',
     'add_bidirectional_mask_if_missing',
